apply TAF folder structure
